[CORE] upgrade version 3.3 vers 3.6 (#3.3_decobois).

+++ htdocs/compta/tva/fiche.php
@@ -1,7 +1,7 @@
 <?php
 /* Copyright (C) 2003      Rodolphe Quiedeville <rodolphe@quiedeville.org>
- * Copyright (C) 2004-2010 Laurent Destailleur  <eldy@users.sourceforge.net>
- * Copyright (C) 2005-2009 Regis Houssin        <regis.houssin@capnetworks.com>
+ * Copyright (C) 2004-2013 Laurent Destailleur  <eldy@users.sourceforge.net>
+ * Copyright (C) 2005-2013 Regis Houssin        <regis.houssin@capnetworks.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -31,59 +31,93 @@ $langs->load("compta");
 $langs->load("banks");
 $langs->load("bills");
 
-$id=$_REQUEST["id"];
-
-$mesg = '';
+$id=GETPOST("id",'int');
+$action=GETPOST('action');
 
 // Security check
 $socid = isset($_GET["socid"])?$_GET["socid"]:'';
 if ($user->societe_id) $socid=$user->societe_id;
 $result = restrictedArea($user, 'tax', '', '', 'charges');
 
+$tva = new Tva($db);
+
 // Initialize technical object to manage hooks of thirdparties. Note that conf->hooks_modules contains array array
-include_once DOL_DOCUMENT_ROOT.'/core/class/hookmanager.class.php';
-$hookmanager=new HookManager($db);
 $hookmanager->initHooks(array('taxvatcard'));
 
 
+
 /**
- * Action ajout paiement tva
+ * Actions
  */
-if ($_POST["action"] == 'add' && $_POST["cancel"] <> $langs->trans("Cancel"))
+
+if ($_POST["cancel"] == $langs->trans("Cancel"))
 {
-    $tva = new Tva($db);
-
-    $db->begin();
-
-    $datev=dol_mktime(12,0,0, $_POST["datevmonth"], $_POST["datevday"], $_POST["datevyear"]);
-    $datep=dol_mktime(12,0,0, $_POST["datepmonth"], $_POST["datepday"], $_POST["datepyear"]);
-
-    $tva->accountid=$_POST["accountid"];
-    $tva->paymenttype=$_POST["paiementtype"];
-    $tva->datev=$datev;
-    $tva->datep=$datep;
-    $tva->amount=$_POST["amount"];
-	$tva->label=$_POST["label"];
-
-    $ret=$tva->addPayment($user);
-    if ($ret > 0)
-    {
-        $db->commit();
-        header("Location: reglement.php");
-        exit;
-    }
-    else
-    {
-        $db->rollback();
-        $mesg='<div class="error">'.$tva->error.'</div>';
-        $_GET["action"]="create";
-    }
+	header("Location: reglement.php");
+	exit;
 }
 
-if ($_GET["action"] == 'delete')
+if ($action == 'add' && $_POST["cancel"] <> $langs->trans("Cancel"))
 {
-    $tva = new Tva($db);
-    $result=$tva->fetch($_GET['id']);
+    $error=0;
+
+	$datev=dol_mktime(12,0,0, $_POST["datevmonth"], $_POST["datevday"], $_POST["datevyear"]);
+    $datep=dol_mktime(12,0,0, $_POST["datepmonth"], $_POST["datepday"], $_POST["datepyear"]);
+
+    $tva->accountid=GETPOST("accountid");
+    $tva->type_payment=GETPOST("type_payment");
+	$tva->num_payment=GETPOST("num_payment");
+    $tva->datev=$datev;
+    $tva->datep=$datep;
+    $tva->amount=GETPOST("amount");
+	$tva->label=GETPOST("label");
+	$tva->note=GETPOST("note");
+	
+	if (empty($tva->datev))
+	{
+		setEventMessage($langs->trans("ErrorFieldRequired",$langs->transnoentitiesnoconv("DateValue")),'errors');
+		$error++;
+	}
+	if (empty($tva->datep))
+	{
+		setEventMessage($langs->trans("ErrorFieldRequired",$langs->transnoentitiesnoconv("DatePayment")),'errors');
+		$error++;
+	}
+	if (empty($tva->type_payment) || $tva->type_payment < 0)
+	{
+		setEventMessage($langs->trans("ErrorFieldRequired",$langs->transnoentitiesnoconv("PaymentMode")),'errors');
+		$error++;
+	}
+	if (empty($tva->amount))
+	{
+		setEventMessage($langs->trans("ErrorFieldRequired",$langs->transnoentitiesnoconv("Amount")),'errors');
+		$error++;
+	}
+
+	if (! $error)
+	{
+		$db->begin();
+
+    	$ret=$tva->addPayment($user);
+		if ($ret > 0)
+		{
+			$db->commit();
+			header("Location: reglement.php");
+			exit;
+		}
+		else
+		{
+			$db->rollback();
+			setEventMessage($tva->error, 'errors');
+			$action="create";
+		}
+	}
+
+	$action='create';
+}
+
+if ($action == 'delete')
+{
+    $result=$tva->fetch($id);
 
 	if ($tva->rappro == 0)
 	{
@@ -96,10 +130,10 @@ if ($_GET["action"] == 'delete')
 			{
 				$accountline=new AccountLine($db);
 				$result=$accountline->fetch($tva->fk_bank);
-				$result=$accountline->delete($user);
+				if ($result > 0) $result=$accountline->delete($user);	// $result may be 0 if not found (when bank entry was deleted manually and fk_bank point to nothing)
 			}
 
-			if ($result > 0)
+			if ($result >= 0)
 			{
 				$db->commit();
 				header("Location: ".DOL_URL_ROOT.'/compta/tva/reglement.php');
@@ -109,18 +143,18 @@ if ($_GET["action"] == 'delete')
 			{
 				$tva->error=$accountline->error;
 				$db->rollback();
-				$mesg='<div class="error">'.$tva->error.'</div>';
+				setEventMessage($tva->error,'errors');
 			}
 	    }
 	    else
 	    {
 	        $db->rollback();
-	        $mesg='<div class="error">'.$tva->error.'</div>';
+	        setEventMessage($tva->error,'errors');
 	    }
 	}
 	else
 	{
-        $mesg='<div class="error">Error try do delete a line linked to a conciliated bank transaction</div>';
+        setEventMessage('Error try do delete a line linked to a conciliated bank transaction','errors');
 	}
 }
 
@@ -145,15 +179,13 @@ if ($id)
 }
 
 // Formulaire saisie tva
-if ($_GET["action"] == 'create')
+if ($action == 'create')
 {
     print "<form name='add' action=\"fiche.php\" method=\"post\">\n";
     print '<input type="hidden" name="token" value="'.$_SESSION['newtoken'].'">';
     print '<input type="hidden" name="action" value="add">';
 
     print_fiche_titre($langs->trans("NewVATPayment"));
-
-    if ($mesg) print $mesg;
 
     print '<table class="border" width="100%">';
 
@@ -178,10 +210,16 @@ if ($_GET["action"] == 'create')
         $form->select_comptes($_POST["accountid"],"accountid",0,"courant=1",1);  // Affiche liste des comptes courant
         print '</td></tr>';
 
-	    print '<tr><td class="fieldrequired">'.$langs->trans("PaymentMode").'</td><td>';
-	    $form->select_types_paiements($_POST["paiementtype"], "paiementtype");
-	    print "</td>\n";
-	    print "</tr>";
+		// Type payment
+		print '<tr><td class="fieldrequired">'.$langs->trans("PaymentMode").'</td><td>';
+		$form->select_types_paiements(GETPOST("type_payment"), "type_payment");
+		print "</td>\n";
+		print "</tr>";
+		
+		// Number
+		print '<tr><td>'.$langs->trans('Numero');
+		print ' <em>('.$langs->trans("ChequeOrTransferNumber").')</em>';
+		print '<td><input name="num_payment" type="text" value="'.GETPOST("num_payment").'"></td></tr>'."\n";
 	}
 
     // Other attributes
@@ -207,8 +245,6 @@ if ($_GET["action"] == 'create')
 
 if ($id)
 {
-    if ($mesg) print $mesg;
-
 	$h = 0;
 	$head[$h][0] = DOL_URL_ROOT.'/compta/tva/fiche.php?id='.$vatpayment->id;
 	$head[$h][1] = $langs->trans('Card');
@@ -224,6 +260,9 @@ if ($id)
 	print '<td width="25%">'.$langs->trans("Ref").'</td><td colspan="3">';
 	print $vatpayment->ref;
 	print '</td></tr>';
+
+	// Label
+	print '<tr><td>'.$langs->trans("Label").'</td><td>'.$vatpayment->label.'</td></tr>';
 
 	print "<tr>";
 	print '<td>'.$langs->trans("DatePayment").'</td><td colspan="3">';
@@ -265,9 +304,20 @@ if ($id)
 	*/
 	print "<div class=\"tabsAction\">\n";
 	if ($vatpayment->rappro == 0)
-		print '<a class="butActionDelete" href="fiche.php?id='.$vatpayment->id.'&action=delete">'.$langs->trans("Delete").'</a>';
+	{
+		if (! empty($user->rights->tax->charges->supprimer))
+		{
+			print '<a class="butActionDelete" href="fiche.php?id='.$vatpayment->id.'&action=delete">'.$langs->trans("Delete").'</a>';
+		}
+		else
+		{
+			print '<a class="butActionRefused" href="#" title="'.(dol_escape_htmltag($langs->trans("NotAllowed"))).'">'.$langs->trans("Delete").'</a>';
+		}
+	}
 	else
+	{
 		print '<a class="butActionRefused" href="#" title="'.$langs->trans("LinkedToAConcialitedTransaction").'">'.$langs->trans("Delete").'</a>';
+	}
 	print "</div>";
 }
 
@@ -275,5 +325,3 @@ if ($id)
 $db->close();
 
 llxFooter();
-
-?>
